+++ SharkScope/Model/Tournaments/TournamentsRequest.swift
@@ -23,7 +23,7 @@ public struct TournamentsRequest: ApiRequest
     
     public var basePath: String { "/poker-statistics/" }
     public var contentType: ContentType { .CSV }
-    public var path: String { "networks/\(network)/players/\(player)/tournaments" }
+    public var path: String { "networks/\(network)/players/\(player)/tournaments.csv" }
     public var parameters: KeyValuePairs<String, String> { [:] }
     public var useCache: Bool = true